12/02/2021 Requirements 17 started

+++ ReCapProject-java-main/src/main/java/com/etiya/RentACar/business/concretes/CityManager.java
@@ -6,11 +6,9 @@ import com.etiya.RentACar.business.dtos.CitySearchListDto;
 import com.etiya.RentACar.business.requests.city.CreateCityRequest;
 import com.etiya.RentACar.business.requests.city.DeleteCityRequest;
 import com.etiya.RentACar.business.requests.city.UpdateCityRequest;
+import com.etiya.RentACar.core.utilities.business.BusinessRules;
 import com.etiya.RentACar.core.utilities.mapping.ModelMapperService;
-import com.etiya.RentACar.core.utilities.results.DataResult;
-import com.etiya.RentACar.core.utilities.results.Result;
-import com.etiya.RentACar.core.utilities.results.SuccessDataResult;
-import com.etiya.RentACar.core.utilities.results.SuccessResult;
+import com.etiya.RentACar.core.utilities.results.*;
 import com.etiya.RentACar.dataAccess.abstracts.CityDao;
 import com.etiya.RentACar.entites.City;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -18,6 +16,7 @@ import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.stream.Collectors;
+
 @Service
 public class CityManager implements CityService {
 
@@ -26,21 +25,25 @@ public class CityManager implements CityService {
     private ModelMapperService modelMapperService;
 
     @Autowired
-    public CityManager(ModelMapperService modelMapperService,CityDao cityDao) {
+    public CityManager(ModelMapperService modelMapperService, CityDao cityDao) {
         this.modelMapperService = modelMapperService;
-        this.cityDao=cityDao;
+        this.cityDao = cityDao;
     }
 
     @Override
     public DataResult<List<CitySearchListDto>> getAll() {
-       List<City> cities = this.cityDao.findAll();
-       List<CitySearchListDto> response = cities.stream().map(city-> modelMapperService.forDto()
-               .map(city , CitySearchListDto.class)).collect(Collectors.toList());
-       return new SuccessDataResult<List<CitySearchListDto>>(response);
+        List<City> cities = this.cityDao.findAll();
+        List<CitySearchListDto> response = cities.stream().map(city -> modelMapperService.forDto()
+                .map(city, CitySearchListDto.class)).collect(Collectors.toList());
+        return new SuccessDataResult<List<CitySearchListDto>>(response);
     }
 
     @Override
     public Result add(CreateCityRequest createCityRequest) {
+        Result result= BusinessRules.run(existsByCityName(createCityRequest.getCityName()));
+        if (result!=null){
+            return result;
+        }
         City city = modelMapperService.forRequest().map(createCityRequest, City.class);
         this.cityDao.save(city);
         return new SuccessResult("City added");
@@ -48,6 +51,10 @@ public class CityManager implements CityService {
 
     @Override
     public Result update(UpdateCityRequest updateCityRequest) {
+        Result result= BusinessRules.run(existsByCityName(updateCityRequest.getCityName()));
+        if (result!=null){
+            return result;
+        }
         City city = modelMapperService.forRequest().map(updateCityRequest, City.class);
         this.cityDao.save(city);
         return new SuccessResult("City updated");
@@ -55,7 +62,28 @@ public class CityManager implements CityService {
 
     @Override
     public Result delete(DeleteCityRequest deleteCityRequest) {
+        Result result= BusinessRules.run(existsByCityId(deleteCityRequest.getCityId()));
+        if (result!=null){
+            return result;
+        }
         this.cityDao.deleteById(deleteCityRequest.getCityId());
         return new SuccessResult("silme işlemi gerçekleşti");
     }
+
+
+    @Override
+    public Result existsByCityId(int cityId) {
+        if (!this.cityDao.existsById(cityId)) {
+            return new ErrorResult("City Bulunamadı");
+        }
+        return new SuccessResult();
+    }
+
+    private Result existsByCityName(String cityName) {
+        if (this.cityDao.existsByCityName(cityName)) {
+            return new ErrorResult("City ismi bulundu başka city giriniz");
+        }
+        return new SuccessResult();
+    }
+
 }
